+++ hydromt_wflow/workflows/__init__.py
@@ -2,6 +2,7 @@
 
 from .basemaps import *
 from .connect import *
+from .demand import *
 from .gauges import *
 from .glaciers import *
 from .landuse import *